requirements.txt
fastapi
uvicorn
redis
aiohttp
pika
requests
pydantic
python-dotenv